Delving Into Initial Modelling Strategies
I read through papers [1] and [2] to look at battery degradation to better understand the frameworks.

To me, [2] was a lot more digestable, and I took some notes stored in "[2] Battery Model" for how we could
model the battery charging based on that paper.

I have not had the chance to read [3] yet, but I figured we could use the papers/ directory for
organization.